feat(spv-793) - simple refactoring
