+++ actions/paging.go
@@ -1,17 +1,18 @@
-package transactions
+package actions
 
 import (
 	"fmt"
 	"github.com/bitcoin-sv/spv-wallet/models"
 	"github.com/gin-gonic/gin"
+	"net/http"
 	"strconv"
 )
 
-func (a *Action) paging(c *gin.Context) {
-	pageFromQueryParam, _ := strconv.Atoi(c.Query("pageFromQueryParam"))
+func Paging(c *gin.Context) {
+	pageFromQueryParam, _ := strconv.Atoi(c.Query("page"))
 	fmt.Println("Page from request", pageFromQueryParam)
 
-	sizeFromQueryParam, _ := strconv.Atoi(c.Query("sizeFromQueryParam"))
+	sizeFromQueryParam, _ := strconv.Atoi(c.Query("size"))
 	fmt.Println("Size from request", sizeFromQueryParam)
 
 	sort := c.QueryArray("sort")
@@ -26,4 +27,8 @@ func (a *Action) paging(c *gin.Context) {
 	fmt.Printf("%+v\n", pageable)
 
 	// call service with pageable data
+
+	c.JSON(http.StatusOK, gin.H{
+		"pageable": "OK",
+	})
 }
